Fixed error in testing files and made Entry entity public
