+++ src/main/java/business/rules/api/APIReader.java
@@ -7,7 +7,7 @@ public interface APIReader{
     public APIDataResponse request(APIDataRequest query, Presenter p);
     String APP_ID = "45a8cbca";
     String APP_KEY = "4343d22a0cc52431a6d07ee2d73a5c46";
-    String HOME = "https://api.edamam.com/api/recipes/v2";
+    String HOME = "https://api.edamam.com/api/recipes/v2";    //The entry point of the API
     String SEARCH_PREFIX = HOME + "?type=public&app_id=" + APP_ID + "&app_key=" + APP_KEY;
     String PLUS = "%2B";
     String AND = "&";
